renaming vennfridge inner dir to app dir
closes issue #67
