+++ app/static/images/diagram.png
