+++ app/templates/diagram.png
